added extra stuff for autos

+++ src/components/ErrorHandler.tsx
@@ -1,13 +1,11 @@
 import { useLayoutEffect } from 'react'
-import { useNavigate } from 'react-router-dom'
 
 export default function ErrorHandler() {
-    const navigate = useNavigate()
-
     useLayoutEffect(() => {
-        alert('There has been an error, re-routing to main page and clearing storage.')
-        localStorage.clear()
-        navigate('/')
+        if (confirm('There has been an error, reload and clear storage?')) {
+            localStorage.clear()
+            window.location.reload()
+        }
     }, [])
 
     return (
